feat: lógica para mensagem no singular com 1 agendamento e tipos corretos em editar agendamentos
