+++ src/components/listaTiposDeConsultas.js
@@ -1,0 +1,11 @@
+export const TiposDeConsulta = [
+    "Consulta de seguimento em Epilepsia",
+    "Consulta de seguimento em Medicina do Sono",
+    "Primeira consulta em Epilepsia",
+    "Primeira consulta em Medicina do Sono",
+    "Primeira consulta em Neurologia Geral",
+    "Retorno de consulta em Epilepsia",
+    "Retorno de consulta em Medicina do Sono",
+    "Retorno de consulta em Neurologia Geral",
+  ];
+
